[listings filter] Added sort_dir_func parameter to listing_column_filters that is used to specify the order of results returned
Moved sort_on and sort_dir from bungeni.models.domain to the descriptors.
Removed sort_replace entirely as the functionality it was intended to provide
is covered by the listing_column and listing_column_filter.

Reenabled sorting in the listing that was temporarily disabled in r9249

Renamed functions in container.py to follow pep8

TO DO: add listing_column_filter for every listing_column in descriptor.py