[security] Elaborating security and permission guidelines (in Bungeni Programming Guide), in particular adding recommendation that UI and Data permissions are kept entirely distinct and independent of each other.
Adding a ui-dedicated "bungeni.ui.View" permission, intended to be used to replace the use of "zope.View" on UI elements (at least those that are not intended to be Public... numerous public UI views are currently protected with "zope.View", requiring that this permission is granted to Anonymous!) so to make possibility the removal of potential interference from intended-to-be-dedicated-to-a-type view permissions that are then possibly redefined to a generic "zope.View" (that are then possibly granted globally).

Globally granting "bungeni.ui.View" to all authenticated users.

Renaming permissions: 
    bungeni.auditlog.View -> bungeni.ui.audit.View
    bungeni.workspace.View -> bungeni.ui.workspace.View    

Deleting unused permission:
    bungeni.TranslateContent