[descriptor] Changing behaviour of MemberURLDisplayWidget to a more general UserURLDisplayWidget that will first try if user is a member of parliament and if so link to the member page (as previous behaviour) and if not will link to the "user directory listing page" (that does not exist yet--so tmp behavior is to link to admin user view if we are admin, else just send unlinked fullname).
Changing UserDelegationDescriptor delegation_id field to also be a "user" field i.e. in "view" mode the delegation user name will also be rendered with UserURLDisplayWidget.

Corrects failures (new in r9753) of when trying to load a membership in "view" mode, for when the user is not a member of parliament e.g. may happen for office or ministry members.