[declarative] Adding implements declaration of custom domain interfaces for newly created custom domain models.
Renamed DocumentDescriptiveProperties to DocDescriptiveProperties.
Fleshing out DocDescriptiveProperties as a DC adapter, and registerering it as base DC adapter for Doc (custom) types. Note: all other custom-type hard-wired DC adapters must go or be made declarataive in config.

Adjusting registration of the "nowediting" UI View to be for instances providing "IFeatureWorkspace" (as opposed to previous "IBungeniParliamentaryContent"), with an assumption that that only (custom) docs that have a "workspace" feature may be thus edited. If assumption is incorrect then either a qualifier in config for whether a custom Doc type is to be considered a parliamentary item may be necessary or (even better) drop/rework this questionable feature altogether.

Correct a typo and a bug in debug.interfaces_iro.