[base form] Reworking ui.forms.common.Baseform to alchemist.ui.BaseForm and moving it up along the inheritance stack to the more logical base spot; making common AddForm/EditForm inherit from only one class (aclhemist.ui AddForm/EditForm), simplifying inheritance, plus removing some duplicated methods (now just inherited). Updating all usage of ui.forms.common.Baseform to use alchemist.ui.BaseForm.
Adjusting definition of capi.schema common.rng SYMBOL_NAME regex to not use "\w" (that seems to not have same implementation across REGEX implementations, but to instead use the spelled out we want this to mean i.e. "a-zA-Z0-9_", thus making it consistent with the equiv regex in naming.is_valid_identifier.