[qualified features] Adding ability to qualify features via specification of <when> scenarios.
Introducing the possibility to add any number of <when> elements, within a <feature> element, with possible discriminators being @subtype and/or @condition. Key aspects of qualified feature syntax and behaviour:
- if set, the @subtype must be a subtype of the feature "subordinate type" e.g. for "event" this could be "assembly_event", "response_event", etc.
- if set, the @condition may be any arbitrary condition that we choose to define in _conditions.py (same as workflow transition conditions).
- at least one of @subtype and @condition must not be None.
- all <when> elements must always follow any <paramater> elements at the feature level, and may only redefine and and all of those same parameters.
- if the @enabled attribute is not specified (as equal to "true), the <when> scenario is disabled, and ignored.
- <when> elements also support a @note attribute, for possible internally useful hints and remarks on this scenario.
- retrieval of a parameter is first attempted on a (the first found, currently) <when> scenario, and if that scenario does not override the parameter then it is retrieved form the fallback/default feature-level parameters, and raising KeyError if that fails.

Updating the RNC schema for workflow for new <when> element.

Adding TEXT_NOT_EMPTY string value pattern, for text fields that may not be empty; using it for parameter.@value, and dropping check for this in application code.

Elaborating Feature.get_param() to now retrieve parameter values from the appropriate scenario. Dropping the internal Feature._p parameter mapping altogether, replacing it with a mapping of feature._When instances, keyed on (@subtype, @condition), with (None, None) being the entry for the fallback/default parameters for the feature.

Adjusting generation of "report_xhtml_template_factory" vocabulary to use qualified "report_tamplates" feature parameter.

Adding sample code for qualified feature usage in chamber "sitting" feature, with the parameter "report_templates" being qualified for the condition "chamber_is_lower_house". Note the @subtype attribute is also available as additional (or only) criterion to qualify a feature.